requirements.txt
selene == 2.0.0rc9
pytest == 8.2.1
allure-pytest
